Split hosted join auth modules for size gate
